requirements.txt
streamlit
requests
newspaper3k
sympy
mistralai
python-dotenv
lxml_html_clean